* FIXED pixelation code for transition should be on MainCamera and keywall should be a prefab
