+++ Assets/PixelateTransitions.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 28ab745f8490b8543ae70f02f8999e2c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
