+++ Assets/Prefabs/KeyWall.prefab.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: aaa3287587a42464c8bd1a2311aacf7b
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 100100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
